// src/lib/generated/org/models/organizationDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * spaces-organizations
 * Kaizen organizations service
 * OpenAPI spec version: 1.0.0
 */
import type { OrganizationDtoSpaceType } from './organizationDtoSpaceType';
import type { OrganizationDtoPriorStatus } from './organizationDtoPriorStatus';
import type { OrganizationStatus } from './organizationStatus';
import type { OrganizationDtoConfig } from './organizationDtoConfig';
import type { OrganizationDtoProject } from './organizationDtoProject';
import type { Country } from './country';
import type { OrganizationType } from './organizationType';
import type { OrganizationDtoGroup } from './organizationDtoGroup';

export interface OrganizationDto {
  /** @nullable */
  modifiedAt: string | null;
  createdAt: string;
  /** @nullable */
  statusLastModifiedAt: string | null;
  /** @nullable */
  statusReason: string | null;
  /** @nullable */
  groupId: number | null;
  /** @nullable */
  projectId: number | null;
  /** @nullable */
  parentId: number | null;
  /** @nullable */
  spaceType: OrganizationDtoSpaceType;
  /** @nullable */
  webUrl: string | null;
  /** @nullable */
  domain: string | null;
  /** @nullable */
  logoUrl: string | null;
  /** @nullable */
  region: string | null;
  /** @nullable */
  city: string | null;
  address: string;
  /** @nullable */
  contactMsisdn: string | null;
  /** @nullable */
  alternateEmail: string | null;
  /** @nullable */
  contactEmail: string | null;
  /** @nullable */
  priorStatus: OrganizationDtoPriorStatus;
  status: OrganizationStatus;
  typeId: number;
  countryId: number;
  /** @nullable */
  code: string | null;
  name: string;
  id: number;
  /** @nullable */
  config?: OrganizationDtoConfig;
  /** @nullable */
  project?: OrganizationDtoProject;
  country?: Country;
  type?: OrganizationType;
  /** @nullable */
  group?: OrganizationDtoGroup;
}